requirements.txt
streamlit
pdf2image
python-dotenv
openai
tiktoken
faiss-cpu
pytesseract
backoff
pathlib
transformers
